Make dependency non generic
This is required as we may have many depedencies with different output
type. We need to be able to abstract over them all so we have a uniform
non generic interface to all rdds.

This was done by splitting the `Rdd` trait into `Rdd` with the non
generic stuff and `TypedRdd`. Too bad Rust doesn't seem to have wildcard
types like java or scala.